backend/requirements.txt
Django==3.1.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
hashids==1.3.1
